Drop Hold State Test
This patch removes the hold state tests. The reasoning is that this
test in unstable and fails all the time in CI, inconveniencing pull
requests authors and maintainers. At the same time, nobody actually
really cares about the hold state any longer and nobody wants to dig
into the test and what's going on to fix this. Therefor, removing this
will not harm functionality and will cause less trouble for the project,
even though it's certainly nice to have more tests.