Let Editor know if files are available locally
This patch is in preparation to fixing #3538: The editor is deployed on
both the admin and the presentation server. But the publication only
points to one of these servers by default. This means that one of the
servers needs to return special HTTP headers so that the other one can
load data or cross-origin requests will fail.

In reality though, most installations will either have the distribution
on a network file system, meaning it is available from all servers and
we could easily circumvent the issue, or they have dedicated and
hopefully properly configured download servers anyway.

In case the latter exist, there shouldn't be an issue in the first
place. This patch now indicates to the editor if the former is the case
and the editor can load files from the local server.

This should always be true in the recommended installation of Opencast,
but this actually lets us react to the cases where it is not true.

The status is indicated in the `/editor/<id>/edit.json`:

```json
{
  "local": true
}
```