Simplify ldap user directory implementation (#4383)
* simplify ldap user directory implementation

Reuse logic of OpencastLdapAuthoritiesPopulator implementation in opencast's user directory.

* remove mixing ldap authorities with internal ones

The OpencastLdapAuthoritiesPopulator returned a mixture of LDAP and
internal authorities. The OpencastLdapAuthoritiesPopulator achieved
this by retrieving the user object via the security service.
This prevents reusing the populator in the LDAP user directory implementation
because it results in a recursion.

* Add UserDetailsContextMapper implementation

Providing the UserDetailsContextMapper as an alternative to the
LdapAuthoritiesMapper offers all opencast roles instead of only roles
derived from LDAP. The OpencastUserDetailsMapper provides a
UserDetails object from Opencast's UserDetailsService, aka
UserAndRoleDirectoryServiceImpl.