Add a module/endpoint/service to redirect users to a given URL
This seems useless at first sight, but enables certain patterns
that are specifically useful for external applications.

An example of this is (pre-)authentication of users
for Opencast services by external applications using a JWT,
without using third party cookies, browsers' local storage,
and also avoiding having the JWT be part of the target URL:
You would have a `form` with `<input type="hidden">` fields
containing the URL of an Opencast service (e.g. Opencast Studio)
and a JWT.
The `action` of the form would be `<opencast>/redirect/get`,
and the method would be `POST`.
On submitting this form, the request would go through the usual
Spring Security mechanisms, giving you a session based on the JWT
(picked up by the `JWTRequestParameterFilter`) and then
redirecting you to the target URL with that session.

This seems a bit roundabout, but I invite everyone to think through
why every piece of this scheme is necessary. Specifically, you can't
use `fetch` (or old style AJAX), because that's not a navigation,
running into the whole third party cookie mine field.