Add rules for merging pull requests to the docs
What a committer is and is not allowed when merging a pull request is not clearly documented. The best way to inform myself about the rules I could currently find is sifting through the proposal logs, which is time consuming, confusing and makes it easy to miss relevant proposals. Therefore, I would like to add the rules in a place in the docs where they are easier to find.