Add a note about protecting the new redirect endpoint(s)
Right now they are protected by default. That is, only admins
can use it. However, for it to be useful, people will want to open
it up a bit more. We don't want to introduce an open redirect vuln
in our default configuration though.

The example (commented out) grants access to JWT-users, and in general
the idea would be to carefully control and dole out the roles
that can use these endpoints so as to avoid people being phished
with links to Opencast.