Catch error when querying non-existing db table
Using a NamedNativeQuery from an unrelated entity seems to not cause JPA to create the table that was queried.
Therefore, the query can cause an error if the table has not been created yet (e.g. because there has
not been an event scheduled yet). This commit catches that error. It might make more sense to relate
the entity "SchedulerIndexData" to "ExtendedEventDto" to create the table, but I am unsure on how
to do that.