Allow hotkeys in create dialogs in input elements
In the Admin UI, the "Create event" and "Create series" dialogs both
have hotkeys with which one can easily jump to previous/next tab of the
dialog (provided the requirements to go to the next tab are met of
course). However, one can get "stuck" in the metadata views if an input
field or dropdown is in focus, as the hotkeys won't trigger in that
case. This fixes that by allowing them to trigger.
Behaviour for any other hotkeys is unchanged.