Change default hotkeys for create dialogs in admin UI
In the admin UI, you can switch between different tabs in a create
dialog via hotkeys. Per default, they use `Enter` and `Backspace`,
which feels weird to use. This commit changes them for the `left` and
`right` arrow keys. It also adds an extra `Shift` as a requirement,
to avoid conflicts with various browsers and operating systems.