Avoid using jobs in SeriesUpdatedEventHandler
When a series has their metadata or ACLs changed, the
SeriesUpdatedEventHandler propagates those changes to all events of the
series. The handler uses lots of Opencast jobs for this, making an
update that should only require a couple milliseconds take a couple
seconds instead.

This tries to make updating events fast by avoiding using jobs for
updating metadata or ACLs, instead calling the code the jobs would run
directly.