Make event updating faster after changes to series metadata/ACL
When there are changes to a series metadata or ACL, all events
belonging to that series need to updated as well. Currently, updating
events takes a very long time.

This commit decreases the time required for updating somewhat by
increasing the polling interval with which Opencast checks if an
updating job has completed. From the default 5 seconds to 100 milli-
seconds.