Treat captions as tracks in standard workflows
Instead of treating captions as something special, this commit attempts
to change the workflows to treat them as tracks, subjecting them to the
same operations.
Actually has to skip some operations in partial-publish that will cause
an exception if given something that is not a media file, but that
should be fine since those operations are irrelevant to subtitles.