Fix reindex of multi-tanant systems
Index of the asset manager fails for multi-tenant systems because it is
assumed that the initiating user (part of one org) has access to events
from other organizations. We already set up the security context with a
system user for the snapshot organization and can reuse that user object
for the reindex as well.