Sets values in specific config file to default
This sets the config values in the listproviders file
`acl.default.create.properties` to their default values. This should
already have happened in the PR associated with the files creation,
but was simply overlooked. With these values set to their defaults,
the ACL tab in the event and series details view in the Admin UI
should function again.