Reimplemented StorageUsage functions for asset storage
As mentioned in the commented, the proper way to do this would be to
return the individual values for each root directory. But as far as I
can tell, that would require rewriting the StorageUsage interface, which
would then require additional changes in the workspace and working file
repository modules, which would be a lot of work for some functions that
only see use in statistics anyway.