Fix documentation of the `language` setting of the Amberscript WOH
"Target language" is actually a term Amberscript uses to denote
the language the transcriptions will end up in. This is, however,
not what `language` specifies. That is rather the language we assume
the recording to be in.

Amberscript allows you to explicitly specify a target language
and will then translate the subtitles to that language. However,
we don't support that setting (yet).