Upload Allinone as Actions Workflow Artifact
We build and test Opencast with each `push` or `pull_request` event on
GitHub actions anyway. But sometimes that is not enough and you will
need to build it locally again to properly test a patch.

This patch will make GitHub store a pre-built artifact along with the
workflow run for the usual 90 days. As a developer, this allows you to
just download and run the tarball for testing, making the review process
easier.