Make Paella 7 Default
This patch updates the default configuration to make Paella 7 the
default player for Opencast 14. Paella 6 is still available though and
adopters can easily switch back by changing a single configuration
option.

This patch also adjust the automatic dependency updates. We argued that
we could do half-tested updates for something that is in beta (and
actually broke the player once or twice in that process), but we cannot
do that for our default players.

Updates should be properly tested from now on and automated updates
should go into `develop`.