Fix REST docs login problem
The REST docs of Opencast have always been publicly available, but
the recently added new admin interface seems to have broken this. Going
to the REST docs overview page and not being logged in causes an empty
page to be displayed.

This patch fixes the issue by moving from using jQuery to implementing
the whole functionality in pure JavaScript which isn't really any harder
with modern JavaScript.