Mitigate not loading ACLs
This patch mitigates #3083.

A race condition between loading organizations and ACLs causes ACL
templates often not being available on fresh installs. The reason for
this is that once ACLs are detected, they are added to all available
organizations. If the organizations are not yet available, no ACLs are
added.

This slight configuration change causes ACLs to be loaded with an
initial delay. Not very elegant, but should fix the issue in basically
all cases.

That said, we should implement a proper solution at one point.