Truncate Bundle Info
This patch add a clean up routine for bundle information from nodes
which no longer exist to the migration scripts.  This is done
automatically for every node if they are shut down *if* they are shut
down properly.  We can safely do this since the Opencast cluster should
be completely shut down during the database migration.