Make Composite Ffmpeg Command Configurable (#4878)
If we want to use GPU's to encode composite videos, we need to change
the part of the ffmpeg command that is currently being constructed by
the workflow operation handler. This makes most of that configurable
(except for the audio mapping), by making use of conditional encoding
profile parameters similarly to the encode operation. It also adds a
bunch of other variables that the workflow operation handler fills
with information needed in the ffmpeg command.

The `composite` encoding profile makes full use of the new configuration
options, while the `mp4-preview.dual` profile was simplified.

Additionally, this also reformats a table for aesthetic purposes only.