Update to Node 18
This patch updates Opencast to use the current LTS version of Node.js
for building the JavaScript components.

This deliberately excludes Paella Player 6 since it's build
infrastructure causes problems. But since we are in the process of
replacing that component anyway, I don't think it is worth working on
that.

This fixes  #4268