Show user information in event workflow details
Currently, if you go to the workflows tab of the event details modal
and check which user started a workflow, you will see that users id.
If you are using auth services like e.g. Shibboleth, the user id
is likely to be a random assortment of letters and numbers that are
not helpful in determining who started the workflow. This PR changes
the "submitter" value returned from the backend, so that instead of
the user id the user name and email are displayed.