Reduce backend requests in events access policy tab
Previously, the frontend would sen a POST request for every user role
in the ACL. Since we get ALL users in another request anyway, this commit
changes that to instead get filter the users from the list of all users
we already have. It does so fairly efficiently by leveraging binary search.