Publish and update maven cache
We have a lot of automated build downloading dependencies from Maven
Central and mvn.opencast.org over and over again. This takes time and
occasionally causes build failures.

This patch adds a workflow building up a repository cache once a week
and making it publically available, so that you can just get most of the
dependencies in just a single tarball. This should make things faster
and easier.