Fix tainted canvas in editor thumbnail extractor
Fixes a CORS issue in the (old) editor. Since 2b92420, the thumbnail
extractor uses a canvas element in the frontend. This canvas element
can become tainted if a proper CORS header is missing on the source,
preventing any image extraction from the canvas. This fixes that by
ensuring that a CORS header is set.