Dont copy media files to new scheduled event
Resolves #4312

In case of adding new media files to a scheduled event through the InboxScanner, Opencast
creates a copy of the scheduled event with the new media files. Unfortunately, this means
it also copies the old media files from the scheduled event, which could lead to
unwanted media files in the new event. Therefore, this commit drops all media files after
copying the mediapackage.

Copying and then dropping the media files, while simple, is arguably a pretty wasteful solution
to the problem. Another approach would be to define which files need to be copied
and then only copy those (i.e. if we actually only need the dublincore, we could simply
create a new mediapackage and copy the dublincore over).