Dont allow to delete user with active workflow in Admin UI
If a user starts a workflow, Opencast expects the user to exist for
at least the duration of that workflow. However, the Admin UI could be used
to delete said user while the workflow was still active, resulting in a
broken workflow. This commit prevents that from happening by checking
whether a user has any active workflows associated with them before allowing
them to be deleted