Allow engage ui to handle non-16/9 thumbnails
The engage ui currently expects that thumbnails for its video
display are 160x90 in size and aspect ratio. This commit allows
for thumbnails of other sizes and ratios to be displayed without
breaking the layout.

The commit is more of a quick fix than the larger rework it would
arguably need to make a well designed thumbnail display for
differing resolutions. But it should be "good enough" for now.