Allow ltitools to handle non-160/90 thumbnails
The ltitools series subtool currently expects that thumbnails
for its video display are exactly 160x90 in size. This commit allows
for thumbnails of other sizes and ratios to be displayed without
breaking the layout.

The commit is more of a quick fix than the larger rework it would
arguably need to make a well designed thumbnail display for
differing resolutions. But it should be "good enough" for now.