Add whisper-ctranslate2 flags to WhisperEngine.java
`whisper-ctranslate2`(https://github.com/Softcatala/whisper-ctranslate2)
is an OpenAI whisper derivative with the main benefit of being significantly
more performant on CPUs.
It also intentionally comes with the same CLI, so it works with our
current implementation of whisper in `WhisperEngine.java`.

This PR makes mention of `whisper-ctranslate2` in the Opencast docs. It
also adds configuration for two potentially useful flags exclusive to
`whisper-ctranslate2` to `WhisperEngine.cfg`.