Add endpoint to get languages & licenses
Attempts to resolve #5099.

Adds two new endpoints to the external API. They return the
languages/licenses configured in the listproviders.

The endpoint behaviour is fairly straightly copied from their
Admin UI endpoint counterpart in `admin-ng/resources`. The key
difference is that `admin-ng/resources` allows for request for any
provider, whereas this commit created two explicit endpoints for
two explicit providers. The reasoning is that users of the external
api should not have access to every provider, and that most providers
only make sense in the context of the Admin UI anyway. I'd be happy
to hear other opinions on this.