Increase index rebuild logging frequency for batches
This commit enables existing logging functionality to increase
the frequency in which services with high element counts log.

The function `logIndexRebuildProgress()` in `AbstractIndexProducer.java`
expects a batch size to be passed if updates are done in batches.
However, now service did that, so the resulting logging was relatively
sporadic. This commit now makes each service pass their batch size,
resulting in more frequent logging.