Only show Asset Upload for options of correct type
Currently the Asset Upload link in the event details is always shown
because ng-show also considers the source options, even though these are
then not displayed, leading to an empty form with a non-function button.
If we don't have options configured for the actual asset upload, we
shouldn't show the link in the first place.