Copy active inputs between CAs if they have the same set of inputs
For a scheduled event in the admin ui, the capture agent for the event
can be changed in the event details. However, when changing the capture
agent, the input channels for the capture agent will always be unset
(except for the capture agent the event was created with). As most
institutions usually use multiple capture agents of the same brand,
with the same set of input channels, having to manually reactivate
every input after switching capture agents is quite cumbersome.

This PR changes the default behaviour for switching capture agents.
If both capture agents have the same set of input channels, which
inpus are active will now be copied over from the previously selected
capture agent to the newly selected one.