Scope dependency `slf4j-log4j12` to `test` by default
Note that this changes the scope in which a couple of modules
depend on this, namely:

- `execute-impl`
- `execute-workflowoperation`
- `oaipmh-persistence`

where it was (implicitly) `compile` before.
I don't see a reason for this, so it should be fine.

I still added an explicit `<scope>test</scope>`.
Looking at other dependencies that are `test`-scoped
in `dependencyManagement` (well, one other ...)
the convention seems to be to still add it
to the individual modules' `dependencies`.
See also the next commit, though.