Replace `slf4j-log4j12` with `slf4j-reload4j`
This changes literally nothing since the former is redirected
to the latter anyway on the level of Maven starting with
the exact version we are using.

If you look carefully at the build output you can even see
Maven warning about this. So I guess this **does** change something,
namely it removes said warning.