Index archived events correctly for multitenant systems
Since we iterate over multiple organizations for a page of 1000 events
in the AssetManager, the resulting set of events might not be dividable
by the bathc size. Since the most recent patch, these events would no
longer be skipped, but they would be indexed for the wrong organization,
namely the one coming after, if there is one.

Also it imo doesn't make any sense to potentially iterate over
organizations for every page, but rather it should be the other way
around.