Fix broken JAR (zip) files
This ensures that our JAR files are actually valid zip files since we
seem to have broken files headers in several releases and this is
causing issues with recent JVM updates.

This is not a very nice fix, but it should do the trick. It seems like
Opencast 15 no longer suffers from this bug. We can probably remove this
workaround in that branch. Though I would maybe keep the integrity test
around for a bit longer.