Fix editing custom actions in the ACL editor
Right now you can't edit custom actions of individual policies
in the ACL editor; all the actions of all policies are kind of
tied together.

This is due to a bug in the JavaScript caused by the select component
all operating on the same shared array. This PR fixes that bug
by just making copies in appropriate places to break that link.