Fix Admin Interface Redirect
With the introduction of the new admin interface, users are dynamically
redirected after the login. Unfortunately, if a user does not have the
necessary permissions, instead of the forbidden/switch user page he used
to end up, there will just be a blank page. That's confusing.

This patch slightly modifies the code so that users end up on the 401
error page after all.