Add migration notes for subtitles as tracks (#4844)
* Add migration notes for subtitles as tracks

Adds some release notes about subtitles as tracks and how to migrate.

Also adds a new WOH for migration. It allows to change the type of
mediapackages, e.g. from attachment to track.

* Addendum to subtitle migration release notes - republish

Indicate that subtitles should be republished after converting them
to tracks from attachments.

* Update docs/guides/admin/docs/releasenotes/subtitles-as-first-class-citizens.txt

Co-authored-by: Miguel Escriva <miesgre@gmail.com>

* Fix wording in changetype-woh.md

Wrongly said something about cloning, which is not what is happening.

* Note valid types in changetype-woh.md

In theory there are more valid types, but these are the ones we
want to support.

* Fix checkstyle in Apereo header

* Replace 14 with 15

Changes didn't make it into 14, so now 15 is the new goal

---------

Co-authored-by: Miguel Escriva <miesgre@gmail.com>