Fixed Admin UI redirect after login
In Opencast 14 the new Admin UI was added as an experimental preview. The opencat administrator should be able to configure, whether the loged in users should be redirected to the new or the previous Admin UI. Due to the issue the users where always redirected to the new Admin UI. This patch fixes it.