Use context aware logger for workflow service
The workflow service is using a special logger that supports units of
work so that you can identify several logger statements belonging to a
specific task.

However, with the new implementation, there is but one unit of work
definition left, and it contains just a single logging statement.
Nothing you actually still need this for.

```
logger.startUnitOfWork();
logger.warn(failureToFail, "Unable to update workflow to failed state");
logger.endUnitOfWork();
```

The downside to using this special logger is that you loose context
information. With the normal logger you get information about the
location where this has been logged. With the special logger this
information is lost. That is really annoying when debugging based on
logs:

```
2023-09-05 18:14:57,449 | INFO  | (ServiceRegistryJpaImpl:622) - Jobs with IDs '[8579450]' deleted
2023-09-05 18:14:57,450 | WARN  | (Log:213) - [>2ef22a08] No events for workflow instance 8579450 found in the Elasticsearch index.
2023-09-05 18:14:57,496 | DEBUG | (Log:213) - [>2ef22a08] Deleting workflow instance 8579614
2023-09-05 18:14:57,548 | INFO  | (Log:213) - [>2ef22a08] Removing temporary files for workflow 8579614
2023-09-05 18:14:57,548 | DEBUG | (Log:213) - [>2ef22a08] Removing temporary file https://oc-admin.virtuos.uni-osnabrueck.de/files/mediapackage/6a092362-56c4-46a9-b82d-f389d97ca19c/security-policy-series/xacml.xml for workflow Workflow {8579614}
```

Given that it is doing more harm than good, this patch switches the
workflow service to the default logger. With this you also now get
proper context information about logs:

```
2023-09-05 18:43:34,069 | INFO  | (ServiceRegistryJpaImpl:622) - Jobs with IDs '[9256488]' deleted
2023-09-05 18:43:34,073 | WARN  | (WorkflowServiceImpl:2260) - No events for workflow instance 9256488 found in the Elasticsearch index.
2023-09-05 18:43:34,082 | DEBUG | (WorkflowServiceImpl:2127) - Deleting workflow instance 9256489
2023-09-05 18:43:34,107 | INFO  | (WorkflowServiceImpl:881) - Removing temporary files for workflow 9256489
```