Check if storage directories are subdirectories of each other
When configuring multiple storage directories, one might accidentally
configure a storage directory that is a subdirectory of another storage
directory. While this *could* work, it would force us to support a lot of
annoying edge cases and it is unlikely that a user would intentionally set
up his storage directories that way. So we disallow it by throwing an exception.