Fix custom roles pattern now working in access policy tab
Intends to fix #5016.

The issue was that the user part of the access policy tab in the details
view in the admin ui would not allow for custom role patterns. With this
PR, you should now able to add custom roles in the user part.

Note that you still have to respect your custom role pattern. So if your
pattern is something like `ROLE_USER_[A-Z][A-Z0-9]*`, you still have
to type the `ROLE_USER_` part. I suppose that could be improved upon,
but I decided it's not worth the effort.