Remove Internal Logger
This patch follows the work started with pull request #5301 and
completely fixes #5261 by removing the internal logger.

The only thing still in use from the logger was a static method
converting seconds to a human readable time string. This patch moves a
simplified version of this method to Opencast's `DateTimeSupport` where
this time-related functionality makes more sense.

Doing that, the `Log` class was then completely unused and could be
removed.