Update Editor to 2023-10-05
Based on the branch https://github.com/opencast/opencast-editor/tree/release-for-oc-14.5
which includes a revert regarding subtitles. The revert is necessary to
work with OC14.

It would be great if someone could test if everything still works, or if
the revert mentioned above (or something that should have been reverted
but wasn't) broke anything.

Known issues not related to the revert:
- https://github.com/opencast/opencast-editor/issues/1154
- https://github.com/opencast/opencast-editor/issues/1161