Update Editor to 2023-10-17
The biggest (and most obvious) change is the graphical redesign, akin to #5318.
Furthermore, this also makes the editor work with the changes to subtitles
coming with Opencast 15.

This goes into develop as most agreed that the redesign is too big of a change to release it as a minor version.

Known major issues (may or may not be related to this release):
- opencast/opencast-editor#1154
- opencast/opencast-editor#1161