Transform language name to ISO code for Whisper
When Whisper uses language detection, it returns the language name in
English instead of the ISO 639 code for some reason. But we need the
latter for the language tag of the track. This makes an attempt to
translate the language name in English to a matching two-letter language
code using locales.