Docs: Hide bot PRs in changelog
When looking through the changelog in the docs to see
what has changed, most people will not care about the
semi-automatic dependency updates. Instead, they pose
a hindrance as they need to be mentally filtered out.
This PR moves all dependabot PRs to the bottom of the
version they appear in, and hides them away in a html
dropdown.