Fix manually entering time in `datetimepicker`-s in Safari
When you try to manually edit the time of a `datetimepicker`
(for example trying to change the start date/time metadata field
of an event), instead of using the combo boxes from the picker,
Safari does weird things to your focus, which makes the task
really difficult.

This is a [known issue](https://github.com/trentrichardson/jQuery-Timepicker-Addon/issues/903)
in the (seemingly unmaintained) `jqueryui-timepicker-addon` library.
Luckily this patch seems to work around it.

It's a bit of a hack, but the thing we are tampering with here
is related to a feature we are not using, so I don't see the harm. 🤷‍♀️

See also https://trentrichardson.com/examples/timepicker/#alt_examples
in the tab "Options" and "Examples" to (try to) understand
what the option means (you might also need
https://api.jqueryui.com/datepicker/#option-altField)
and so verify that the bug is not Opencast, but in the library already.