Add search index sort multi field
The default sorting of the search index could be more intuitive. For
example, lowercase letters follow uppercase letters, which leads to
funny results that by sorting ascending, the letter `a` comes after `Z`.

By introducing a multi-field `sort` for selected fields, it is
possible to normalize the string to lowercase, and with asciifolding,
better results can be achieved.

An even better approach would be to use `icu_collation` filter, which
can be configured to use a collation table for a specific language.

Signed-off-by: Gregor Eichelberger <gregor.eichelberger@tuwien.ac.at>