Update the default admin interface
This patch switches to the new admin interface as the default interface
for Opencast.

Users can still switch back to the old interface, but this move should
push adoption, ensure that people actually test the new interface and
force people to develop features against the new, not the old interface.