(Re-)Enable the JWT feature by default
The feature was initially enabled by default
by virtue of it not actually being a (Karaf) feature,
but having its bundles be part of `opencast-core` directly.
It was later disabled by default in #3065
because of some flaky boot issues observed by Innsbruck.
That's where the Karaf feature comes from,
which you can install and enable in your site configuration
if you are not affected by that issue,
as most adopters weren't.
In #3173 we added some documentation to that effect
and re-included the feature in the distributions,
so that you only need to enable it.

Since we haven't heard anyone else complain so far
and JWT is become more important,
we decided to give it another go and just enable it again.
In that spirit, this PR reverts the documentation parts of #3173,
and adds the feature created by #3065 to the default boot features
of all the distribtuions.

I opted not to just revert #3065 because I felt there is a shift
towards a more modular Opencast in the community
which I'm generally supportive of.
There was some talk in Matrix hinting at the preference
for this feature to be a plugin,
the plugin manager itself is an independent Karaf feature,
there is also the CAS feature, etc.

Unfortunately, the JWT feature currently can't be a plugin
because it contains a fragment bundle hosted by `kernel`,
so if it were to be activated by the plugin manager after boot,
`kernel` would have to restart, which Opencast doesn't survive atm.
But we can honour the idea of more modularity by keeping the feature
instead of just reintegrating the bundles into `opencast-core`.

Plus this has the added benefit of being rather easily deactivated
if need be:
Just do the reverse of what the docs that this PR throws out say
and remove the feature from the boot features of your Karaf config.