Fix Docs Landing Page
This patch fixes the docs landing page which has  a broken HTML (and
arguably also bootstrap) syntax: Missing closing tags and too many
columns for them to be actually rendered.

Additionally, this changes the calendar link from an ICS file you have
to import into your calendar to the web view. The reason for that change
is that as a new user, I might want to take a look, but I don't
necessarily want to immediately import the whole calender into my own
calender.