Whisper transcription engine bug fixing
This patch fixes a number of minor bugs and inconveniences in the
whisper transcription engine. This includes loading of default values
when the configuration has been modified which incorrectly kept their
old values in some cases. It also does proper debug logging, not
filtering out the actual output, which is what you want for debugging.