Don't set bogus default email for admin
This patch updates the default value of the admin user's email address
so that the user has no address at all instead of providing a bogus one
which will usually not work.

This allows users, for example, the use the `send-email` operation ans
skip over the admin instead of failing on the (usually) invalid address.

- This succeeds #5310
- This is based on #5476