Docs: Add beginner guide for submitting issues
We have detailed information on how to deal with security issues,
but no general guide for submitting issues. This aims to change that.
The goal here is to add helpful information for people who have been
told to "submit an issue", but only have a vague idea on what that
entails and are inexperienced with the Github web interface.