Docs: Make Developer Overview page more verbose
There's but a single sentence of introduction on the landing page
of the developer docs. This aims to change that, so users do not
feel quite as lost.

This also tries to provide a loose definition for the different sections,
based on what should be in those sections, not what is actually in them.