Add support for custom actions in ACL to Tobira harvest API
Previously, only read and write roles were transmitted. I would slightly
prefer having a map from roles to "list of actions", but the map from
actions to "list of roles" is how the API currently works so in order
to preserve backwards compatibility, we just keep it. It doesn't
matter anyway.

The search service only contains `read` and `write` actions, so for 
events, we have to check the media package which we do via the
AuthorizationService.