Fix ACL template selection breaking after first selection
The template selection dropdown for access policies in e.g. the
event details is broken. After selecting an access policy, when trying
to select another access policy, the rules from the access policy above
the selected access policy will be selected.
For example, if you have `public, authenticated, private` and select
`public`, you get the aces for `public`. If you then select `authenticated`
you get the aces for `public`. Then selecting `private` will lead to an
error.

This fixes that. Apparently the angularjs select thingy breaks if you
set ng-model to a value that is not in ng-options (e.g. setting it
ng-model to `0`, but ng-options only contains `951, 952, 953`).
This commit sets it to `undefined` instead, which angularjs seems to be
able to handle.