Hint at how to properly close issues with pull requests
Mentioning issues in your PR with some magic keywords can
automatically close these issues when the PR is merged. We even hint
at this in our PR-template already.

However, GitHubs implementation of this does not work super well with
our release/branching model. Specifically, it only works if the PR is
merged into the default branch, which is develop in our case.

It can be made to work in our situation if the magic phrases are
mentioned in commit messages alternatively or in addition to the PR
description, because we regularly forward merge the release branches.

This patch adds a corresponding note about this to the documentation,
and links it in the PR template.